Add a `/random` endpoint
Previously, we could view a deterministic avatar at `/:seed` or get
redirected to a random `/:seed` page from the root (`/`). To test new
functionality, we would go to `/`, get redirected, view the generated
avatar, and then type in the URL bar again to go back to `/` and restart
the process. This is tedious.

To fix this, I've added a `/random` endpoint that generates a random
avatar and displays it in place. This means we can just keep hitting
refresh to generate a new avatar.

It is important to keep the `/random` action above `/:seed` in the
`app.rb` file, otherwise the router things that `/random` is `/:seed`
where the seed is the string "random". In the future, maybe it makes
sense to namespace `/:seed` under some value such as `/avatars/:seed`.